+++ ospf6d/ospf6_intra.h
@@ -55,30 +55,6 @@ extern in_addr_t conf_debug_ospf6_brouter_specific_area_id;
 	(IS_OSPF6_DEBUG_BROUTER_SPECIFIC_AREA                                  \
 	 && conf_debug_ospf6_brouter_specific_area_id == (area_id))
 
-/* Router-LSA */
-#define OSPF6_ROUTER_LSA_MIN_SIZE              4U
-struct ospf6_router_lsa {
-	uint8_t bits;
-	uint8_t options[3];
-	/* followed by ospf6_router_lsdesc(s) */
-};
-
-/* Link State Description in Router-LSA */
-#define OSPF6_ROUTER_LSDESC_FIX_SIZE          16U
-struct ospf6_router_lsdesc {
-	uint8_t type;
-	uint8_t reserved;
-	uint16_t metric; /* output cost */
-	uint32_t interface_id;
-	uint32_t neighbor_interface_id;
-	in_addr_t neighbor_router_id;
-};
-
-#define OSPF6_ROUTER_LSDESC_POINTTOPOINT       1
-#define OSPF6_ROUTER_LSDESC_TRANSIT_NETWORK    2
-#define OSPF6_ROUTER_LSDESC_STUB_NETWORK       3
-#define OSPF6_ROUTER_LSDESC_VIRTUAL_LINK       4
-
 enum stub_router_mode {
 	OSPF6_NOT_STUB_ROUTER,
 	OSPF6_IS_STUB_ROUTER,
@@ -98,42 +74,6 @@ enum stub_router_mode {
 	(ntohl(((struct ospf6_router_lsdesc *)(x))->neighbor_interface_id))
 #define ROUTER_LSDESC_GET_NBR_ROUTERID(x)                                      \
 	(((struct ospf6_router_lsdesc *)(x))->neighbor_router_id)
-
-/* Network-LSA */
-#define OSPF6_NETWORK_LSA_MIN_SIZE             4U
-struct ospf6_network_lsa {
-	uint8_t reserved;
-	uint8_t options[3];
-	/* followed by ospf6_netowrk_lsd(s) */
-};
-
-/* Link State Description in Router-LSA */
-#define OSPF6_NETWORK_LSDESC_FIX_SIZE          4U
-struct ospf6_network_lsdesc {
-	in_addr_t router_id;
-};
-#define NETWORK_LSDESC_GET_NBR_ROUTERID(x)                                     \
-	(((struct ospf6_network_lsdesc *)(x))->router_id)
-
-/* Link-LSA */
-#define OSPF6_LINK_LSA_MIN_SIZE               24U /* w/o 1st IPv6 prefix */
-struct ospf6_link_lsa {
-	uint8_t priority;
-	uint8_t options[3];
-	struct in6_addr linklocal_addr;
-	uint32_t prefix_num;
-	/* followed by ospf6 prefix(es) */
-};
-
-/* Intra-Area-Prefix-LSA */
-#define OSPF6_INTRA_PREFIX_LSA_MIN_SIZE       12U /* w/o 1st IPv6 prefix */
-struct ospf6_intra_prefix_lsa {
-	uint16_t prefix_num;
-	uint16_t ref_type;
-	uint32_t ref_id;
-	in_addr_t ref_adv_router;
-	/* followed by ospf6 prefix(es) */
-};
 
 
 #define OSPF6_ROUTER_LSA_SCHEDULE(oa)                                          \
